fix(portfolio): rename actions directory to `_actions`
